Added lap functionality and styling
