+++ src/FormatTime.tsx
@@ -1,0 +1,11 @@
+export default function formatTime(time: number): string {
+  const hours = Math.floor(time / 360000);
+  const minutes = Math.floor((time % 360000) / 6000);
+  const seconds = Math.floor((time % 6000) / 100);
+  const milliseconds = time % 100;
+  return `${hours.toString().padStart(2, "0")}:${minutes
+    .toString()
+    .padStart(2, "0")}:${seconds.toString().padStart(2, "0")}:${milliseconds
+    .toString()
+    .padStart(2, "0")}`;
+}
