+++ src/StopWatch.tsx
@@ -1,5 +1,6 @@
 import React from "react";
-import "./StopWatch.css";
+import "./styles/StopWatch.css";
+import formatTime from "./FormatTime";
 export default function StopWatch(props: {
   isStarted: boolean;
   timer: number;
@@ -9,17 +10,6 @@ export default function StopWatch(props: {
   const timer = props.timer;
   const setTimer = props.setTimer;
 
-  function formatTime(time: number): string {
-    const hours = Math.floor(time / 360000);
-    const minutes = Math.floor((time % 360000) / 6000);
-    const seconds = Math.floor((time % 6000) / 100);
-    const milliseconds = time % 100;
-    return `${hours.toString().padStart(2, "0")}:${minutes
-      .toString()
-      .padStart(2, "0")}:${seconds.toString().padStart(2, "0")}:${milliseconds
-      .toString()
-      .padStart(2, "0")}`;
-  }
   React.useEffect(() => {
     if (isStarted) {
       const interval = setInterval(() => {
