+++ src/styles/Lap.css
@@ -1,0 +1,13 @@
+.lapList{ 
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    gap: 4px;
+    margin-top: 20px;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+    font-size: 62px;
+    font-family: Arial, Helvetica, sans-serif;
+    font-weight: bold;
+}
