+++ src/styles/StopWatchButton.css
